+++ ui/src/components/MewContent.vue
@@ -10,6 +10,7 @@
             @click.stop
           >
             {{ contentPart[0] }}
+            <q-tooltip :delay="TOOLTIP_DELAY">{{ contentPart[1] }}</q-tooltip>
           </a>
 
           <router-link
@@ -30,7 +31,7 @@
 
 <script setup lang="ts">
 import { PATH, ROUTES } from "@/router";
-import { FeedMew, LinkTargetName } from "@/types/types";
+import { FeedMew, LinkTargetName, TOOLTIP_DELAY } from "@/types/types";
 import { TAG_REGEX, TAG_SYMBOLS } from "@/utils/tags";
 import { serializeHash } from "@holochain-open-dev/utils";
 import { computed, PropType } from "vue";