+++ ui/src/components/CreateMewField.vue
@@ -254,9 +254,17 @@ const onLinkTextKeyDown = (keyDownEvent: KeyboardEvent) => {
 
     const anchor = document.createElement("a");
     anchor.href = "#";
-    anchor.textContent = linkText.value
-      ? TAG_SYMBOLS.URL + linkText.value
-      : url;
+
+    // Wrap link text in brackets if it has multiple words
+    const linkTextValue = linkText.value?.trim();
+    if (linkTextValue?.includes(" ")) {
+      anchor.textContent = `${TAG_SYMBOLS.URL}[${linkTextValue}]`;
+    } else if (linkText.value) {
+      anchor.textContent = TAG_SYMBOLS.URL + linkTextValue;
+    } else {
+      anchor.textContent = url;
+    }
+
     anchor.dataset[ANCHOR_DATA_ID_URL] = url ?? undefined;
     range.insertNode(anchor);
     // insert space after link