+++ CHANGELOG.md
@@ -4,6 +4,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
 - Support multi-word link text in mews with the syntax `^[my link]`
+- Human-readable timestamps
+- Search for hashtags, cashtags
+- Display URL in tooltip when hovering over mew links
 
 ### Added
 - URL hyperlinks in mews.