+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/). This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+- Support multi-word link text in mews with the syntax `^[my link]`
 - Human-readable timestamps
 - Search for hashtags, cashtags
 - Display URL in tooltip when hovering over mew links