+++ CHANGELOG.md
@@ -3,14 +3,15 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/). This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+
+### Added
 - Support multi-word link text in mews with the syntax `^[my link]`
 - Human-readable timestamps
 - Search for hashtags, cashtags
 - Display URL in tooltip when hovering over mew links
-
-### Added
 - URL hyperlinks in mews.
 ### Changed
+- **BREAKING CHANGE**: Upgrade to Holochain v0.1.3
 ### Fixed
 ### Removed