+++ CHANGELOG.md
@@ -5,6 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## Unreleased
 
 ### Added
+### Changed
+### Fixed
+### Removed
+
+## 2023-02-26 v0.9.0-alpha
+### Added
 - Support multi-word link text in mews with the syntax `^[my link]`
 - Human-readable timestamps
 - Search for hashtags, cashtags
@@ -12,10 +18,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - URL hyperlinks in mews.
 ### Changed
 - **BREAKING CHANGE**: Upgrade to Holochain v0.1.3
-### Fixed
-### Removed
 
-## 2022-12.25 v0.8.0-alpha
+## 2022-12-25 v0.8.0-alpha
 ### Added
 - Thread view for a mew with all its replies.
 ### Changed